backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
requests
python-multipart

langgraph
langchain
langchain-google-genai
langchain-openai
tavily-python
google-search-results
